address all comments
